+++ alacritty/src/cli.rs
@@ -272,6 +272,7 @@ impl Options {
 
         config.hold = self.hold;
 
+        config.set_dynamic_title(config.dynamic_title() && self.title.is_none());
         config.window.dimensions = self.dimensions.unwrap_or(config.window.dimensions);
         config.window.title = self.title.unwrap_or(config.window.title);
         config.window.position = self.position.or(config.window.position);
@@ -284,8 +285,6 @@ impl Options {
                 config.window.class.general = general.into();
             }
         }
-
-        config.set_dynamic_title(config.dynamic_title() && config.window.title == DEFAULT_NAME);
 
         config.debug.print_events = self.print_events || config.debug.print_events;
         config.debug.log_level = max(config.debug.log_level, self.log_level);
@@ -328,12 +327,12 @@ mod tests {
     }
 
     #[test]
-    fn dynamic_title_overridden_by_config() {
+    fn dynamic_title_not_overridden_by_config() {
         let mut config = Config::default();
 
         config.window.title = "foo".to_owned();
         let config = Options::default().into_config(config);
 
-        assert!(!config.dynamic_title());
+        assert!(config.dynamic_title());
     }
 }